pyproject.toml
[tool.poetry]
name = "resume-pdf-generator"
version = "0.1.0"
description = ""
authors = ["Bhanwar <bhanwar.bps89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fpdf2 = "^2.8.1"
markdown2 = "^2.5.1"
streamlit = "^1.40.0"
pyperclip = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
